[UPDATE] Added more test for 100% coverage

+++ app/Http/Controllers/EventQuestionController.php
@@ -10,6 +10,10 @@ class EventQuestionController extends Controller
 {
     public function store(Request $request, Event $event)
     {
+        if (auth()->user()->role !== 'organiser') {
+            abort(403);
+        }
+
         $request->validate(['question' => 'required|string|max:255|min:5']);
 
         EventQuestion::create([
@@ -22,7 +26,11 @@ class EventQuestionController extends Controller
 
     public function destroy($id)
     {
-        $question = EventQuestion::findOrFail($id); // returns 404 if not found
+        $question = EventQuestion::findOrFail($id);
+
+        if (auth()->user()->role !== 'organiser') {
+            abort(403);
+        }
 
         $eventId = $question->event_id;
         $question->delete();
